+++ Projects/Carat/ISCaratMonitor.h
@@ -10,8 +10,8 @@ public:
 	static ISCaratMonitor& Instance();
 
 public:
-	void IncomingQuerySize(quint64 QuerySize); //Регистрация размера входящего запроса
-	void AnswerQuerySize(quint64 AnswerSize); //Регистрация размера ответа на запрос
+	void IncomingQuerySize(int QuerySize); //Регистрация размера входящего запроса
+	void AnswerQuerySize(int AnswerSize); //Регистрация размера ответа на запрос
 
 private:
 	ISCaratMonitor();
@@ -22,8 +22,8 @@ private:
 	ISCaratMonitor& operator=(ISCaratMonitor&&) = delete;
 
 private:
-	quint64 _IncomingQuerySize; //Размер входящего траффика
-	quint64 _AnswerQuerySize; //Размер исходящего траффика
+	int _IncomingQuerySize; //Размер входящего траффика
+	int _AnswerQuerySize; //Размер исходящего траффика
 };
 //-----------------------------------------------------------------------------
 #endif
